Began work on DP. Up to filling DDI_BUF_TRANS
